ci_framework/tests/requirements.txt
pytest-mock>=3.11.1
